+++ lib/lightning/policies/project_users.ex
@@ -14,8 +14,8 @@ defmodule Lightning.Policies.ProjectUsers do
           | :rerun_job
           | :create_job
           | :delete_job
-          | :delete_project
           | :access_project
+          | :delete_project
           | :create_workflow
           | :edit_project_name
           | :edit_digest_alerts
@@ -39,7 +39,12 @@ defmodule Lightning.Policies.ProjectUsers do
           Lightning.Projects.Project.t()
         ) :: boolean
   def authorize(:access_project, %User{} = user, %Project{} = project),
-    do: Projects.is_member_of?(project, user)
+    do:
+      is_nil(project.scheduled_deletion) and
+        Projects.is_member_of?(project, user)
+
+  def authorize(:delete_project, %User{} = user, %Project{} = project),
+    do: Projects.get_project_user_role(user, project) == :owner
 
   def authorize(
         action,
@@ -48,9 +53,6 @@ defmodule Lightning.Policies.ProjectUsers do
       )
       when action in [:edit_digest_alerts, :edit_failure_alerts],
       do: id == user_id
-
-  def authorize(:delete_project, %User{} = user, %Project{} = project),
-    do: Projects.get_project_user_role(user, project) in [:owner]
 
   def authorize(action, %User{} = user, %Project{} = project)
       when action in [