+++ lib/lightning_web/router.ex
@@ -92,6 +92,7 @@ defmodule LightningWeb.Router do
       live "/settings/projects", ProjectLive.Index, :index
       live "/settings/projects/new", ProjectLive.Index, :new
       live "/settings/projects/:id", ProjectLive.Index, :edit
+      live "/settings/projects/:id/delete", ProjectLive.Index, :delete
 
       live "/settings/audit", AuditLive.Index, :index
 
@@ -104,6 +105,7 @@ defmodule LightningWeb.Router do
         live "/jobs", JobLive.Index, :index
 
         live "/settings", ProjectLive.Settings, :index
+        live "/settings/delete", ProjectLive.Settings, :delete
 
         live "/runs", RunLive.Index, :index
         live "/runs/:id", RunLive.Show, :show